QcomModulePkg: Add support to the updated hyp_boot_info struct
Remove deprecated fields and add new fields in the hyp_boot_info
structure. The new hyp_boot_info struct has multiple dtbos
that need to be overlaid with the primary VM dtb. So update the
hyp_boot_info struct and add the required support accordingly.

Adding this to support VMs on SA410M.

Change-Id: Ie32ddb7951bfcd2423d99c2b543b08d3de7480e8