QcomModulePkg: Remove offset pointer optimization in inftrees.c.
inftrees.c was subtracting an offset from a pointer to an array,
in order to provide a pointer that allowed indexing starting at
the offset. This is not compliant with the C standard, for which
the behavior of a pointer decremented before its allocated memory
is undefined. Per the recommendation of a security audit of the
zlib code by Trail of Bits and TrustInSoft, in support of the
Mozilla Foundation, this tiny optimization was removed, in order
to avoid the possibility of undefined behavior.

Change-Id: Ic4e1543e1ce2a04023286896bbfee8b394131c9d
Git-commit: 6a043145ca6e9c55184013841a67b2fef87e44c0
Git-repo: https://github.com/madler/zlib
Signed-off-by: Vishnu Teja <quic_vteja@quicinc.com>