libavb: Split boot image to load and verify in parallel.
To reduce the boot time, it is required to divide boot image
load and verification in to two parts which can work
independently.
Boot image is divided in to two parts. Load of first part and
second part can be done by first cpu.
Verification of first part and second part can be done
by second cpu.
Hence verification of first part will go in parallel with
load of second part.

Change-Id: I40a910a21b1bfd8911253880efc3660a7cd4f082
Signed-off-by: Santosh Dronamraju <quic_sdronamr@quicinc.com>