Hibernation: Update resume kernel parameter and Validate swap slot before restore
Dynamically figure out the swap partition
system path and append to the kernel command line.

If MMU uses the swap due to memory pressure, the  swap_map_page is
not at disk offset 1. In this case, the swsusp_header->image
points to the first swap_map_page, from which restore
is to be done.

Use macro for code readability and to make independent of a/b slot
enable.

Partition Error checking is done during every read_image which is
unnecessary. Move error check out of read_image.

At times kernel memory management use swap partition.
Restore code is not handling this scenario. So abort
hibernation in such cases.

Change-Id: I973a7e99f7906bcd5e6f69a62a04cc02992bfeb3