libavb： load and verify of hash image using multithreading
To reduce the load and verification of the hash image time
like the boot image, the image is split into several chunks,
the chunk size is mainly determined by the image size and is
calculated by an approximate optimal formula.   chunkSize=
sqrt(|READ_SPEED| * |IO_COST_TIME|* |ImageSize|) / |PageSize| * |PageSize|)

Load Thread and Verify Thread are on different cpu. Semaphores
are used for synchronization, In the first stage, load thread
loads specified chunksize and post semaphore SemLoadFirst
sequentially.verify thread waits for SemLoadFirst to verify
chunk sequentially.In the second stage, which means the last
chunk is going to be load and verified. Once the the semaphore
SemLoadSecond is posted, verify thread will finish the verification
work.

Change-Id: I9775ce86dc538d4d963201999c994c67ce46c878
Signed-off-by: Wei Liu <quic_wliu8@quicinc.com>