QcomModulePkg: Fix issue gzip header extra field with inflate()
If the extra field was larger than the space the user provided with
inflateGetHeader(), and if multiple calls of inflate() delivered
the extra header data, then there could be a buffer overflow of the
provided space. This commit assures that provided space is not
exceeded.

Change-Id: Ia2f9444119ead906efb59b64a66a090578868e39
Git-repo: https://github.com/madler/zlib
Git-commit: eff308af425b67093bab25f80f1ae950166bece1
Signed-off-by: Vishnu Teja <quic_vteja@quicinc.com>