edk2: QcomModulePkg: fix for device boot-up issue at fastboot.
The device crashes when the fastboot boot boot.img command is executed,
while the device is in fastboot mode due to unbound condition in string search.
NULL is added at the end of the array of strings to fix this issue.

Test: device booted with fastboot boot boot.img command
Change-Id: I2d36b4ab442e6c177db598103ad1081737c09a94
(cherry picked from commit 2937a12d23497575426e9550fa8cc70548346b60)