QcomModulePkg: Bootlib: Add node name length check when looking up a node
For device tree node searching, the full node name will be stored into a
list, which may contain name, alias and address. However, the search
only find the keywords, such as 'soc', if there is a similar name 'soc1',
it may return the offset for 'soc' to 'soc1'. Therefore, add a check for
node length.

Change-Id: I620cc94876e9f25a490be4b2153fa9cb4aaf2130