QcomModulePkg: Bootlib: Add partition name length check when switching slots
When switching slots, if only a certain length of the partition name is
matched, it can cause slot switching errors. For example, the partition
information of xbl_a might be incorrectly switched to xbl_ac_config,
causing the device to fail to boot. Therefore, a check for the length
of the partition name has been added.

Change-Id: I9ee9f833e01c4fd615a83463b8e63141f48d052c