Support OCTET STRING attribute values in X509_NAME
Cherry-pick from
https://boringssl-review.googlesource.com/c/boringssl/+/78327/4
and will be superceded as part of the next BoringSSL rollup
in http://ag/32943738 but CP-ing separately in order to enable
further cherry-picking to P25 branches.

This is part of a broader issue with OpenSSL's goofy ASN1_PRINTABLE type
(which seems to have no connection to any standard), but this is a
minimal fix intended for easy cherry-picking.

This now leaves ASN1_PRINTABLE unused, but for now I've forked it into
an ASN1_ANY_AS_STRING type. It is still very far from an actual ANY, but
the intent is to capture AttributeValue's status as an ANY type, but
which OpenSSL's API forces us to represent as an ASN1_STRING. (Later
changes will make it actually match the name.)

Update-Note: X.509 name attributes now may be OCTET STRINGs.

Bug: 42290275
Bug: 409165331
Test: TH
Change-Id: I1389533595a972bd0b4aa3c1840dc0f15c0fa645