Revert "QcomModulePkg: Add SoftSku Id in DT selection logic"
This reverts commit 10638ec276613faac5c1f19789272b3c7a4bbd20.

In earlier implementation : bits 26-29 of msm-id field were reused for softsku-id. But later it was decided to not use these bits as it will limit the number of FPs supported in feature thereby making it non-scalable. Also, Feature Pack(FP) provides more of software details rather than HW details of device. Hence, separate softsku-id field is added in DT.

Change-Id: I9378b289365c83263ade23256bbd18433915b334