QcomModulePkg: Add Qtvm Dtbo partition authentication
Before this change, the qtvm_dtbo image partition was not signed.
With this, the partition's authenticity is verified and the QTVM won't be loaded
if the partition's hash footer created by avbtool.py doesn't match the content's
digest.

A distinct AuthQtvmDtboImg() function was created which is heavily based off of
the LoadImageAndAuthVB2() function above it. The reason for a distinct function is
that the authentication of the qtvm_dtbo partition does not rely on placing its auth
hash in the vbmeta partition; its hash footer is placed at the end of the partition.
This is leveraged by using the flag AVB_SLOT_VERIFY_FLAGS_NO_VBMETA_PARTITION when
calling avb_slot_verify().

For this to work, the flashed qtvm_dtb.img needs to have been signed by the private
key with avbtool.py

Change-Id: I2b08e66ca64a798602f4ede4f2049e738dcc3a7e
Signed-off-by: Vivekanand Tryambake <vtryamba@qti.qualcomm.com>