QcomModulePkg: Ensure 4KB Alignment for PVM Firmware Size
When the bootloader is unlocked, the PVM firmware size is set to the
partition size. However, when the device is locked, the size is set to
the image size. This results in error in the mem donate RM call
when the image size is not aligned to 4KB.

Fixed this issue by always using the image size with 4KB alignment.

Change-Id: Icc3405c5208f30d83d73133ed9e5c329298a6af0
Signed-off-by: Sreenad Menon <quic_sreemeno@quicinc.com>